# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST para gestionar múltiples sesiones de Telegram via MTProto",
        "title": "Telegram API",
        "contact": {},
        "version": "0.3.0"
    },
    "host": "localhost:7789",
    "basePath": "/api/v1",
    "paths": {
...
        "/sessions/{id}/qr/regenerate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Genera un nuevo código QR para una sesión existente que falló o expiró",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sessions"
                ],
                "summary": "Regenerar QR de sesión",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
        "/sessions/{id}/submit-password": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Completa autenticación con 2FA password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sessions"
                ],
                "summary": "Submit 2FA Password",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "2FA Password",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.SubmitPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.TelegramSession"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request - INVALID_PASSWORD",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict - ALREADY_AUTHENTICATED",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/sessions/import-tdata": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Importa sesión desde Telegram Desktop",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sessions"
                ],
                "summary": "Import TData Session",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "API ID from my.telegram.org",
                        "name": "api_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API Hash from my.telegram.org",
                        "name": "api_hash",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Session name (optional)",
                        "name": "session_name",
                        "in": "formData",
                        "required": false
                    },
                    {
                        "type": "file",
                        "description": "TData files (key_data.json, etc.)",
                        "name": "tdata",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.TelegramSession"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request - VALIDATION",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity - TDATA_INVALID",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/sessions/{id}/webhook": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna configuración actual del webhook",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhooks"
                ],
                "summary": "Obtener webhook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.WebhookConfig"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Configura URL de webhook para recibir eventos de la sesión",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhooks"
                ],
                "summary": "Configurar webhook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Configuración",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.WebhookCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.WebhookResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Elimina configuración de webhook",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhooks"
                ],
                "summary": "Eliminar webhook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.AuthMethod": {
            "type": "string",
            "enum": [
                "sms",
                "qr",
                "tdata"
            ],
            "x-enum-varnames": [
                "AuthMethodSMS",
                "AuthMethodQR",
                "AuthMethodTData"
            ]
        },
...
        "domain.SubmitPasswordRequest": {
            "type": "object",
            "required": [
                "password"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "description": "2FA password"
                }
            }
        },
...
        "domain.SessionStatus": {
            "type": "string",
            "enum": [
                "pending",
                "code_sent",
                "password_required",
                "authenticated",
                "failed",
                "banned",
                "frozen"
            ],
            "x-enum-varnames": [
                "SessionPending",
                "SessionCodeSent",
                "SessionPasswordRequired",
                "SessionAuthenticated",
                "SessionFailed",
                "SessionBanned",
                "SessionFrozen"
            ]
        },
        "domain.TelegramSession": {
            "type": "object",
            "properties": {
                "api_id": {
                    "type": "integer"
                },
                "auth_method": {
                    "$ref": "#/definitions/domain.AuthMethod"
                },
                "auth_state": {
                    "$ref": "#/definitions/domain.SessionStatus"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "phone_number": {
                    "type": "string"
                },
                "session_name": {
                    "type": "string"
                },
                "telegram_user_id": {
                    "type": "integer"
                },
                "telegram_username": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "domain.WebhookConfig": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "events": {
                    "description": "Tipos de eventos a enviar",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "last_error": {
                    "type": "string"
                },
                "last_error_at": {
                    "type": "string"
                },
                "max_retries": {
                    "type": "integer"
                },
                "secret": {
                    "description": "Para firmar requests",
                    "type": "string"
                },
                "session_id": {
                    "type": "string"
                },
                "timeout_ms": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "domain.WebhookCreateRequest": {
            "type": "object",
            "required": [
                "url"
            ],
            "properties": {
                "events": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "message.new",
                        "message.edit"
                    ]
                },
                "max_retries": {
                    "type": "integer",
                    "example": 3
                },
                "secret": {
                    "type": "string",
                    "example": "mi_secret_123"
                },
                "timeout_ms": {
                    "type": "integer",
                    "example": 5000
                },
                "url": {
                    "type": "string",
                    "example": "https://mi-servidor.com/webhook"
                }
            }
        },
        "domain.WebhookResponse": {
            "type": "object",
            "properties": {
                "events": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "session_id": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "handler.MetaResponse": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "handler.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/handler.ErrorResponse"
                },
                "meta": {
                    "$ref": "#/definitions/handler.MetaResponse"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}