# internal/handler/session_handler_tdata.go
package handler

import (
	"io"
	"mime/multipart"
	"strconv"

	"telegram-api/internal/domain"
	"telegram-api/internal/middleware"
	"telegram-api/pkg/logger"

	"github.com/gofiber/fiber/v2"
)

// validateTDataForm validates the tdata import form fields.
func validateTDataForm(c *fiber.Ctx) (userID string, apiID int, apiHash, sessionName string, err error) {
	userIDUUID, err := middleware.GetUserID(c)
	if err != nil {
		return "", 0, "", "", err
	}

	apiID, err = strconv.Atoi(c.FormValue("api_id", "0"))
	if err != nil || apiID <= 0 {
		return "", 0, "", "", fiber.NewError(400, "api_id is required and must be positive")
	}

	apiHash = c.FormValue("api_hash", "")
	if apiHash == "" {
		return "", 0, "", "", fiber.NewError(400, "api_hash is required")
	}

	sessionName = c.FormValue("session_name", "")
	return userIDUUID.String(), apiID, apiHash, sessionName, nil
}

// readTDataFiles reads all uploaded tdata files.
func readTDataFiles(files []*multipart.FileHeader) (map[string][]byte, error) {
	tdataFiles := make(map[string][]byte)

	for _, fileHeader := range files {
		file, err := fileHeader.Open()
		if err != nil {
			logger.Error().Err(err).Str("filename", fileHeader.Filename).Msg("Error opening tdata file")
			return nil, fiber.NewError(400, "Error reading tdata file")
		}
		defer func() {
			if err := file.Close(); err != nil {
				logger.Error().Err(err).Str("filename", fileHeader.Filename).Msg("Error closing tdata file")
			}
		}()

		content, err := io.ReadAll(file)
		if err != nil {
			logger.Error().Err(err).Str("filename", fileHeader.Filename).Msg("Error reading tdata file content")
			return nil, fiber.NewError(400, "Error reading tdata file content")
		}

		tdataFiles[fileHeader.Filename] = content
	}

	return tdataFiles, nil
}

// buildTDataResponse builds the response for tdata import.
func buildTDataResponse(session *domain.TelegramSession) fiber.Map {
	return fiber.Map{
		"session": fiber.Map{
			"session_id":       session.ID,
			"is_active":        session.IsActive,
			"telegram_user_id": session.TelegramUserID,
			"username":         session.TelegramUsername,
			"auth_state":       session.AuthState,
			"auth_method":      "tdata",
		},
	}
}

// ImportTData imports Telegram Desktop session
// @Summary Import Telegram Desktop session
// @Description Import Telegram Desktop session from tdata files
// @Tags Sessions
// @Accept multipart/form-data
// @Produce json
// @Security BearerAuth
// @Param api_id formData int true "API ID"
// @Param api_hash formData string true "API Hash"
// @Param session_name formData string false "Session Name"
// @Param files formData file true "TData files"
// @Success 201 {object} handler.Response{data=fiber.Map}
// @Failure 400 {object} handler.Response
// @Failure 422 {object} handler.Response
// @Router /sessions/import-tdata [post].
func (h *SessionHandler) ImportTData(c *fiber.Ctx) error {
	userID, err := middleware.GetUserID(c)
	if err != nil {
		return c.Status(401).JSON(NewErrorResponse(401, "Not authenticated"))
	}

	form, err := c.MultipartForm()
	if err != nil {
		logger.Error().Err(err).Msg("Error parsing multipart form")
		return c.Status(400).JSON(NewErrorResponse(400, "Invalid multipart form"))
	}

	userIDStr, apiID, apiHash, sessionName, err := validateTDataForm(c)
	if err != nil {
		return c.Status(400).JSON(NewErrorResponse(400, err.Error()))
	}

	files := form.File["files"]
	if len(files) == 0 {
		return c.Status(400).JSON(NewErrorResponse(400, "tdata files are required"))
	}

	tdataFiles, err := readTDataFiles(files)
	if err != nil {
		return c.Status(400).JSON(NewErrorResponse(400, err.Error()))
	}

	logger.Debug().
		Str("user_id", userIDStr).
		Int("api_id", apiID).
		Str("session_name", sessionName).
		Int("files_count", len(tdataFiles)).
		Msg("Processing tdata import request...")

	session, err := h.service.ImportTData(c.Context(), userID, apiID, apiHash, sessionName, tdataFiles)
	if err != nil {
		return handleSessionError(c, err)
	}

	response := buildTDataResponse(session)
	return c.Status(201).JSON(NewSuccessResponse(response))
}